+++ clinicadl/config/config/maps_manager.py
@@ -31,17 +31,3 @@ class MapsManagerConfig(BaseModel):
             raise ClinicaDLArgumentError(
                 "Cannot save nifti if the network task is not reconstruction. Please remove --save_nifti option."
             )
-
-    def adapt_config_with_maps_manager_info(self, maps_manager: MapsManager):
-        if not self.split_list:
-            self.split_list = maps_manager._find_splits()
-        logger.debug(f"List of splits {self.split_list}")
-
-        if self.diagnoses is None or len(self.diagnoses) == 0:
-            self.diagnoses = maps_manager.diagnoses
-
-        if not self.batch_size:
-            self.batch_size = maps_manager.batch_size
-
-        if not self.n_proc:
-            self.n_proc = maps_manager.n_proc
