+++ clinicadl/train/tasks/regression/regression_cli_options.py
@@ -26,7 +26,7 @@ selection_metrics = cli_param.option_group.task_group.option(
     "--selection_metrics",
     "-sm",
     multiple=True,
-    type=click.Choice(get_type("selection_metrics", regression_config)),
+    type=get_type("selection_metrics", regression_config),
     default=get_default("selection_metrics", regression_config),
     help="""Allow to save a list of models based on their selection metric. Default will
     only save the best model selected on loss.""",
@@ -35,7 +35,7 @@ selection_metrics = cli_param.option_group.task_group.option(
 loss = cli_param.option_group.task_group.option(
     "--loss",
     "-l",
-    type=click.Choice(get_type("loss", regression_config)),
+    type=get_type("loss", regression_config),
     default=get_default("loss", regression_config),
     help="Loss used by the network to optimize its training task.",
     show_default=True,
